refactor(tests): reorganize into folder-per-module / file-per-function
Split monolithic per-module test files into a nested structure with one
folder per source file/class and one file per function or method. Shared
helpers (_FakeSession/_FakeResponse/patch_aiohttp and write_pyproject)
are moved into folder-level conftest.py files.
