+++ tests/test_environment_handler/test_delete_venv.py
@@ -1,0 +1,31 @@
+from pytest_mock import MockerFixture
+
+from uv_audit.environment_handler import EnvironmentHandler
+
+
+def test_delete_venv_when_folder_exists(mocker: MockerFixture):
+    # arrange
+    handler = EnvironmentHandler()
+    mocker.patch("uv_audit.environment_handler.os.path.exists", return_value=True)
+    rmtree = mocker.patch("uv_audit.environment_handler.shutil.rmtree")
+
+    # act
+    result = handler.delete_venv()
+
+    # assert
+    assert result is True
+    rmtree.assert_called_once_with(handler._folder)
+
+
+def test_delete_venv_when_folder_does_not_exist(mocker: MockerFixture):
+    # arrange
+    handler = EnvironmentHandler()
+    mocker.patch("uv_audit.environment_handler.os.path.exists", return_value=False)
+    rmtree = mocker.patch("uv_audit.environment_handler.shutil.rmtree")
+
+    # act
+    result = handler.delete_venv()
+
+    # assert
+    assert result is True
+    rmtree.assert_not_called()
