+++ tests/test_environment_handler/test_list_packages.py
@@ -1,0 +1,30 @@
+from pytest_mock import MockerFixture
+
+from uv_audit.environment_handler import EnvironmentHandler
+
+
+def test_list_packages_returns_empty_list_when_no_output(mocker: MockerFixture):
+    # arrange
+    handler = EnvironmentHandler()
+    mocker.patch.object(handler, "run_command", return_value="")
+
+    # act
+    result = handler.list_packages()
+
+    # assert
+    assert result == []
+
+
+def test_list_packages_returns_parsed_packages(mocker: MockerFixture):
+    # arrange
+    handler = EnvironmentHandler()
+    pip_output = (
+        "Package    Version\n---------- -------\nclick      8.2.1\nrequests   2.32.3\n"
+    )
+    mocker.patch.object(handler, "run_command", return_value=pip_output)
+
+    # act
+    result = handler.list_packages()
+
+    # assert
+    assert result == ["click==8.2.1", "requests==2.32.3"]
