--- tests/test_file_handler.py
@@ -1,172 +1,0 @@
-from pathlib import Path
-
-import pytest
-from pytest_mock import MockerFixture
-
-from uv_audit.file_handler import _report_vulns, handle_file, handle_pyproject
-from uv_audit.pyproject_handler import PyProjectSelection
-
-
-def test_handle_pyproject_runs_full_flow(mocker: MockerFixture, tmp_path: Path):
-    # arrange
-    pyproject = tmp_path / "pyproject.toml"
-    pyproject.write_text("[project]\nname='x'\nversion='0.0.0'\ndependencies=[]\n")
-    selection = PyProjectSelection(
-        path=pyproject,
-        extras=[],
-        groups=[],
-        has_main_deps=True,
-    )
-
-    env_cls = mocker.patch("uv_audit.file_handler.EnvironmentHandler")
-    env = env_cls.return_value
-    env.list_packages.return_value = ["requests==2.32.3"]
-
-    scanner_cls = mocker.patch("uv_audit.file_handler.VulnerabilityScanner")
-    scanner_cls.return_value.run_check.return_value = []
-
-    # act
-    vulns = handle_pyproject(selection)
-
-    # assert
-    env.create_venv.assert_called_once()
-    env.install_pyproject.assert_called_once_with(selection)
-    env.list_packages.assert_called_once()
-    env.delete_venv.assert_called_once()
-    assert vulns == []
-
-
-def test_handle_pyproject_skips_install_when_nothing_selected(
-    mocker: MockerFixture, tmp_path: Path
-):
-    # arrange
-    pyproject = tmp_path / "pyproject.toml"
-    pyproject.write_text("[project]\nname='x'\nversion='0.0.0'\n")
-    selection = PyProjectSelection(
-        path=pyproject,
-        extras=[],
-        groups=[],
-        has_main_deps=False,
-    )
-
-    env_cls = mocker.patch("uv_audit.file_handler.EnvironmentHandler")
-    env = env_cls.return_value
-    env.list_packages.return_value = []
-
-    # act
-    vulns = handle_pyproject(selection)
-
-    # assert
-    env.install_pyproject.assert_not_called()
-    assert vulns == []
-
-
-def test_handle_pyproject_installs_when_extras_only(
-    mocker: MockerFixture, tmp_path: Path
-):
-    # arrange
-    pyproject = tmp_path / "pyproject.toml"
-    pyproject.write_text("[project]\nname='x'\nversion='0.0.0'\n")
-    selection = PyProjectSelection(
-        path=pyproject,
-        extras=["cli"],
-        groups=[],
-        has_main_deps=False,
-    )
-
-    env_cls = mocker.patch("uv_audit.file_handler.EnvironmentHandler")
-    env = env_cls.return_value
-    env.list_packages.return_value = []
-
-    # act
-    handle_pyproject(selection)
-
-    # assert
-    env.install_pyproject.assert_called_once_with(selection)
-
-
-def test_report_vulns_prints_summary_when_vulns_found(
-    capsys: pytest.CaptureFixture[str],
-):
-    # arrange
-    results = [
-        {
-            "package": "flask",
-            "version": "1.1.2",
-            "vulnerabilities": [
-                {
-                    "id": "GHSA-XYZ",
-                    "fixed_in": ["2.0.0"],
-                    "link": "https://example.com",
-                }
-            ],
-        }
-    ]
-
-    # act
-    vulns = _report_vulns(results)
-
-    # assert
-    captured = capsys.readouterr()
-    assert len(vulns) == 1
-    assert vulns[0]["Name"] == "flask"
-    assert "Found 1 known vulnerabilities" in captured.out
-    assert "flask" in captured.out
-    assert "GHSA-XYZ" in captured.out
-
-
-def test_report_vulns_prints_no_vulnerabilities_when_empty(
-    capsys: pytest.CaptureFixture[str],
-):
-    # act
-    vulns = _report_vulns([])
-
-    # assert
-    captured = capsys.readouterr()
-    assert vulns == []
-    assert "No known vulnerabilities found" in captured.out
-
-
-def test_handle_file_runs_full_flow_and_returns_vulns(mocker: MockerFixture):
-    # arrange
-    env_cls = mocker.patch("uv_audit.file_handler.EnvironmentHandler")
-    env = env_cls.return_value
-    env.list_packages.return_value = ["flask==1.1.2"]
-
-    scanner_cls = mocker.patch("uv_audit.file_handler.VulnerabilityScanner")
-    scanner_cls.return_value.run_check.return_value = [
-        {
-            "package": "flask",
-            "version": "1.1.2",
-            "vulnerabilities": [{"id": "GHSA-1", "fixed_in": ["2.0.0"], "link": "x"}],
-        }
-    ]
-
-    # act
-    vulns = handle_file(file_path="/req.txt", is_file=True)
-
-    # assert
-    env.create_venv.assert_called_once()
-    env.install_requirements.assert_called_once_with(
-        requirements_file="/req.txt", is_file=True
-    )
-    env.delete_venv.assert_called_once()
-    assert len(vulns) == 1
-
-
-def test_handle_file_returns_empty_when_no_vulns(mocker: MockerFixture):
-    # arrange
-    env_cls = mocker.patch("uv_audit.file_handler.EnvironmentHandler")
-    env = env_cls.return_value
-    env.list_packages.return_value = ["x==1"]
-
-    scanner_cls = mocker.patch("uv_audit.file_handler.VulnerabilityScanner")
-    scanner_cls.return_value.run_check.return_value = [
-        {"package": "x", "version": "1", "vulnerabilities": []}
-    ]
-
-    # act
-    vulns = handle_file(file_path="/req.txt", is_file=True)
-
-    # assert
-    assert vulns == []
