+++ tests/test_pyproject_handler/test_resolve_selection.py
