+++ tests/test_vulnerability_scanner/conftest.py
@@ -1,0 +1,41 @@
+from typing import Any
+
+from pytest_mock import MockerFixture
+
+
+class _FakeResponse:
+    def __init__(self, status: int, payload: dict[str, Any]):
+        self.status = status
+        self._payload = payload
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, *args):
+        return False
+
+    async def json(self):
+        return self._payload
+
+
+class _FakeSession:
+    def __init__(self, responses: dict[str, _FakeResponse]):
+        self._responses = responses
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, *args):
+        return False
+
+    def get(self, url: str, timeout=None):
+        return self._responses[url]
+
+
+def patch_aiohttp(mocker: MockerFixture, responses: dict[str, _FakeResponse]):
+    mocker.patch(
+        "uv_audit.vulnerability_scanner.aiohttp.ClientSession",
+        return_value=_FakeSession(responses),
+    )
+    mocker.patch("uv_audit.vulnerability_scanner.aiohttp.TCPConnector")
+    mocker.patch("uv_audit.vulnerability_scanner.aiohttp.ClientTimeout")
