+++ tests/test_vulnerability_scanner/test_get_package_url.py
@@ -1,0 +1,12 @@
+from uv_audit.vulnerability_scanner import VulnerabilityScanner
+
+
+def test_get_package_url_format():
+    # arrange
+    scanner = VulnerabilityScanner()
+
+    # act
+    url = scanner._get_package_url("requests", "2.32.3")
+
+    # assert
+    assert url == "https://pypi.org/pypi/requests/2.32.3/json"
