+++ tests/test_vulnerability_scanner/test_run_check.py
@@ -1,58 +1,9 @@
-from typing import Any
-
 import pytest
 from pytest_mock import MockerFixture
 
 from uv_audit.vulnerability_scanner import VulnerabilityScanner
 
-
-class _FakeResponse:
-    def __init__(self, status: int, payload: dict[str, Any]):
-        self.status = status
-        self._payload = payload
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, *args):
-        return False
-
-    async def json(self):
-        return self._payload
-
-
-class _FakeSession:
-    def __init__(self, responses: dict[str, _FakeResponse]):
-        self._responses = responses
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, *args):
-        return False
-
-    def get(self, url: str, timeout=None):
-        return self._responses[url]
-
-
-def _patch_aiohttp(mocker: MockerFixture, responses: dict[str, _FakeResponse]):
-    mocker.patch(
-        "uv_audit.vulnerability_scanner.aiohttp.ClientSession",
-        return_value=_FakeSession(responses),
-    )
-    mocker.patch("uv_audit.vulnerability_scanner.aiohttp.TCPConnector")
-    mocker.patch("uv_audit.vulnerability_scanner.aiohttp.ClientTimeout")
-
-
-def test_get_package_url_format():
-    # arrange
-    scanner = VulnerabilityScanner()
-
-    # act
-    url = scanner._get_package_url("requests", "2.32.3")
-
-    # assert
-    assert url == "https://pypi.org/pypi/requests/2.32.3/json"
+from .conftest import _FakeResponse, patch_aiohttp
 
 
 def test_run_check_returns_vulnerabilities_filtered_by_withdrawn(
@@ -67,7 +18,7 @@ def test_run_check_returns_vulnerabilities_filtered_by_withdrawn(
             {"id": "VULN-2", "withdrawn": "2024-01-01", "details": "retracted"},
         ]
     }
-    _patch_aiohttp(mocker, {url: _FakeResponse(200, payload)})
+    patch_aiohttp(mocker, {url: _FakeResponse(200, payload)})
 
     # act
     results = scanner.run_check(["requests==2.32.3"])
@@ -85,7 +36,7 @@ def test_run_check_returns_empty_when_no_vulnerabilities(mocker: MockerFixture):
     scanner = VulnerabilityScanner()
     url = "https://pypi.org/pypi/requests/2.32.3/json"
     payload = {"vulnerabilities": []}
-    _patch_aiohttp(mocker, {url: _FakeResponse(200, payload)})
+    patch_aiohttp(mocker, {url: _FakeResponse(200, payload)})
 
     # act
     results = scanner.run_check(["requests==2.32.3"])
@@ -100,7 +51,7 @@ def test_run_check_returns_empty_when_status_not_200(mocker: MockerFixture):
     # arrange
     scanner = VulnerabilityScanner()
     url = "https://pypi.org/pypi/requests/2.32.3/json"
-    _patch_aiohttp(mocker, {url: _FakeResponse(404, {})})
+    patch_aiohttp(mocker, {url: _FakeResponse(404, {})})
 
     # act
     results = scanner.run_check(["requests==2.32.3"])
@@ -114,7 +65,7 @@ def test_run_check_returns_empty_when_status_not_200(mocker: MockerFixture):
 def test_run_check_raises_on_empty_requirements(mocker: MockerFixture):
     # arrange
     scanner = VulnerabilityScanner()
-    _patch_aiohttp(mocker, {})
+    patch_aiohttp(mocker, {})
 
     # act / assert
     with pytest.raises(ValueError, match=r"No valid packages found in requirements\."):
@@ -124,7 +75,7 @@ def test_run_check_raises_on_empty_requirements(mocker: MockerFixture):
 def test_run_check_ignores_malformed_requirements(mocker: MockerFixture):
     # arrange
     scanner = VulnerabilityScanner()
-    _patch_aiohttp(mocker, {})
+    patch_aiohttp(mocker, {})
 
     # act / assert
     with pytest.raises(ValueError, match=r"No valid packages found in requirements\."):
@@ -136,7 +87,7 @@ def test_run_check_with_multiple_packages(mocker: MockerFixture):
     scanner = VulnerabilityScanner()
     url_requests = "https://pypi.org/pypi/requests/2.32.3/json"
     url_flask = "https://pypi.org/pypi/flask/1.1.2/json"
-    _patch_aiohttp(
+    patch_aiohttp(
         mocker,
         {
             url_requests: _FakeResponse(200, {"vulnerabilities": []}),
@@ -157,7 +108,7 @@ def test_run_check_strips_whitespace_in_requirements(mocker: MockerFixture):
     # arrange
     scanner = VulnerabilityScanner()
     url = "https://pypi.org/pypi/requests/2.32.3/json"
-    _patch_aiohttp(mocker, {url: _FakeResponse(200, {"vulnerabilities": []})})
+    patch_aiohttp(mocker, {url: _FakeResponse(200, {"vulnerabilities": []})})
 
     # act
     results = scanner.run_check(["  requests == 2.32.3  "])
